Rendering Mode Rework (currently windows only)
Someone recreated JS Engine's render mode with LUA... and this is a port of that code to source! (definitely not taken from SGWLC's engine!!)
This version of Rendering Mode is much faster than before, and doesn't need you to render the song to an MP4 after it's finished, because now the game does it for you!