a big commit Number 2
The functionality of the "JS Engine Recharts" option was refactored
Fixed a random Week 4 crash (hopefully, closing #721)
Added a 5 second wait for update.bat. Hopefully fixes the problem with JSE not updating!
The engine can now originally tell if you played a song in Freeplay and had the "Always Play Cutscenes" option enabled!
The Rating Popup spawning was reworked a bit, and now no longer causes lag when too many popups are spawned at once
WEEKEND 1 IS HERE.

Current issues with Weekend 1:
The camera isn't positioned correctly in the Darnell cutscene
If playing in Story Mode, the game crashes after completing 2hot. I've found that the issue occurs when loading the post-cutscene video, but have no idea how to fix it.
The same thing occurs if you turn on "Always Play Cutscenes" and complete either 2hot or Blazin'.